session_supervisor: carve dispatch.rs (pure move)
